requirements.txt
Flask==3.0.0
gunicorn==21.2.0
supabase==2.13.0
google-generativeai==0.8.3
websockets>=14.0
cachetools==5.3.3
python-dotenv==1.0.1
requests==2.31.0
pytz==2024.1
Authlib==1.3.0
Flask-Limiter==3.7.0
